WIP: implement kernel timestamps for received packets on Linux
This is a proof-of-concept of using kernel timestamps to improve the
accuracy of RTT estimation in Babel.

From quick tests on a local veth pair, it seems to improve accuracy by
around 100 µs on each side (which means a ~200 µs improvement in RTT
accuracy if both babeld speakers use kernel timestamps).

There are several issues:

- kernel timestamps use the realtime clock, and it's not configurable.
  This is really annoying because babeld uses the monotonic clock (for
  good reasons).  Using kernel timestamps forces us to fall back to the
  realtime clock elsewhere in babeld, and it would require some work to do
  it cleanly.

- kernel timestamps are only used for received packets.  The sending side
  (timestamp in Hello) still uses userspace timestamps, with the ensuing
  accuracy issue.  It does not seem possible to tell the kernel to embed a
  timestamp at a specified location in a packet just before sending it,
  unless maybe playing with eBPF.